feat(dts): make any 'any' optional
currently 
`function(optionalParameter)` is interpreted as `function(optionalParameter: any)`
since nothing is known about the optionalParameter, it might have been optional and if it is, the resulting d.ts isn't working (as it is for our spoonx plugins). 
thus, i suggest, to treat any unknown parameter as optional. so:
`function(optionalParameter)` is interpreted as `function(optionalParameter?: any)`